day04/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
matplotlib>=3.7.0
plotly>=5.17.0
pandas>=2.0.0
pytest>=7.4.0
biopython>=1.81
